// src/app/core/models/masters/project-management/project-time-category.ts
export interface ProjectTimeCategory {
    id: number;
    title: string;
    description:string;
    status: string;
    status_id :number;
    status_label: string;
}

export interface ProjectTimeCategoryPaginationResponse {
    current_page: number;
    from:number;
    total: number;
    per_page: number;
    last_page: number;
    data: ProjectTimeCategory[];
}